Minimize daemon telemetry changes
Share telemetry test sinks and remove unrelated formatting and cleanup from the per-server telemetry diff.


Copilot-Session: c5d4f77e-69f4-4116-aa63-b7c1ab0822cb
