+++ .github/memory/FILE_MAP.md
@@ -22,7 +22,7 @@ This file is a **top-level map only**. For per-area directory detail, read the m
 | `Workspaces/` | ide | Solution/Project/Document model, MSBuild loading, Remote (OOP). |
 | `Features/`, `EditorFeatures/` | ide | IDE feature logic and editor integration. |
 | `Analyzers/`, `CodeStyle/` | ide | IDE0xxx code-style analyzers & fixes. |
-| `LanguageServer/` | ide | LSP protocol, server executable, shared daemon identity/telemetry helpers, daemon/thin-client hosting, and LSP tests. |
+| `LanguageServer/` | ide | LSP server. |
 | `VisualStudio/` | ide | VS language services, UI, and EditorConfig template/wizard/command packaging under `EditorConfig/`. |
 | `Razor/src/` | razor | Razor compiler + tooling (own sub-tree layout). |
 | `Scripting/`, `Interactive/` | — | C#/VB scripting engine and REPL. |
