+++ src/LanguageServer/Microsoft.CodeAnalysis.LanguageServer.UnitTests/Microsoft.CodeAnalysis.LanguageServer.UnitTests.csproj
@@ -27,7 +27,9 @@
     <InternalsVisibleTo Include="IdeCoreBenchmarks" />
   </ItemGroup>
 
-  <ItemGroup Label="Shared named-pipe helper for daemon tests">
+  <ItemGroup Label="Shared named-pipe helper for daemon client tests">
+    <!-- Tests connect to the daemon as a client using the same compiler-server pipe helper the thin client
+         uses, so the test's client stream agrees with the daemon's server stream on pipe path -->
     <Compile Include="..\..\Compilers\Shared\NamedPipeUtil.cs" Link="Daemon\NamedPipeUtil.cs" />
   </ItemGroup>
