+++ src/LanguageServer/Protocol/LspServices/LspServices.cs
@@ -27,7 +27,7 @@ internal sealed class LspServices : ILspServices, IMethodHandlerProvider
     /// so these are manually created in <see cref="RoslynLanguageServer"/>.
     /// </summary>
     private readonly FrozenDictionary<string, ImmutableArray<BaseService>> _baseServices;
-    private readonly RoslynTelemetry? _telemetry;
+    private readonly RoslynTelemetry _telemetry = RoslynTelemetry.Current;
 
     /// <summary>
     /// Gates access to <see cref="_servicesToDispose"/> and <see cref="_servicesToDisposeAsync"/>.
@@ -63,13 +63,6 @@ internal sealed class LspServices : ILspServices, IMethodHandlerProvider
         _lazyMefLspServices = serviceMap.ToFrozenDictionary();
 
         _baseServices = baseServices;
-        var telemetryTypeName = typeof(RoslynTelemetry).FullName;
-        Contract.ThrowIfNull(telemetryTypeName);
-        if (_baseServices.TryGetValue(telemetryTypeName, out var telemetryServices) &&
-            telemetryServices is [var telemetryService])
-        {
-            _telemetry = (RoslynTelemetry)telemetryService.GetInstance(this);
-        }
 
         void AddSpecificService(Lazy<ILspService, LspServiceMetadataView> serviceGetter)
         {
@@ -178,7 +171,7 @@ internal sealed class LspServices : ILspServices, IMethodHandlerProvider
             // A service can first be requested from a context that carries no ambient instance of its own (for
             // example a file-watcher callback or work-queue batch), so re-establish this server's instance for
             // factories that capture RoslynTelemetry.Current.
-            using var _ = _telemetry is null ? null : RoslynTelemetry.SetCurrent(_telemetry);
+            using var _ = RoslynTelemetry.SetCurrent(_telemetry);
             var lspService = lazyService.Value;
             if (checkDisposal)
             {
